Use f32 samples
Instead of accepting i16s and casting to f32s, just use f32, and let client
code adapt to it if necessary.

It might be cool to make this whole thing generic over the sample type, but I'm
not sure what needs to happen to work with integer vs. floating point data
(apart from just casting to a float, as this lib used to do). Maybe it could be
generic over floating point types.

My motivation for this change is that JACK processing callbacks get samples as
windows of f32 data anyway, so this change saves me converting to/from i16s.